requirements.txt
alembic==1.8.1
Flask-Migrate==2.1.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.1
flask-restx==0.5.1
psycopg2-binary==2.9.3
python-dotenv==0.21.0
SQLAlchemy==1.4.41
werkzeug==2.1.2
gunicorn==20.1.0
